FS20G03-8 #comment display player count (in progress) + create & join game in redux
